requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
attrs==25.3.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.0
faiss-cpu==1.12.0
fastapi==0.117.1
fonttools==4.60.0
frozenlist==1.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
imbalanced-learn==0.12.4
imblearn==0.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.3.27
langchain-community==0.3.30
langchain-core==0.3.77
langchain-openai==0.3.34
langchain-text-splitters==0.3.11
langsmith==0.4.32
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.9.4
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.0.2
openai==2.1.0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
watchdog==6.0.0
watchfiles==1.1.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
